terminado

+++ src/app/dashboard/products/page.js
@@ -4,6 +4,7 @@ import ProductTable from "@/components/productsComponents/productsTable";
 import { IoAdd } from "react-icons/io5";
 import ProductAddModal from "@/components/productsComponents/addProduct";
 import { useStore } from "@/stores/autenticacion";
+import Swal from "sweetalert2";
 
 export default function ProductsPage() {
   const [isAddModalOpen, setIsAddModalOpen] = useState(false);
@@ -13,7 +14,7 @@ export default function ProductsPage() {
 
   useEffect(() => {
     traerProducto();
-  },[] );
+  }, []);
 
   const traerProducto = async () => {
     try {
@@ -46,10 +47,17 @@ export default function ProductsPage() {
         }),
       });
       const data = await res.json();
-  
+
       if (res.ok) {
         await traerProducto();
         setIsAddModalOpen(false);
+        await Swal.fire({
+          position: "center",
+          icon: "success",
+          title: "Producto agregado",
+          showConfirmButton: false,
+          timer: 1500,
+        });
       } else {
         console.error("Error al agregar el producto:", data.message);
       }
@@ -58,36 +66,43 @@ export default function ProductsPage() {
     }
   };
 
-  const handleUpdateProduct = async (updatedProduct) => {
+  const handleUpdateProduct = async (formData) => {
     try {
-      const res = await fetch(`http://localhost:3010/producto/${updatedProduct.id}`, {
+      const datos = { nombre: formData.nombre, precio: Number(formData.precio), categoria: formData.categoria }
+      console.log(datos)
+      const res = await fetch(`http://localhost:3010/producto/${formData.id}`, {
         method: "PATCH",
         headers: {
-          Authorization: `Bearer ${user.token}`,
           "Content-Type": "application/json",
+          'Authorization': `Bearer ${user.token}`,
         },
-        body: JSON.stringify(updatedProduct),
-      });
-
-      const data = await res.json();
-      console.log(updatedProduct);
+        body: JSON.stringify(datos),
+      })
+      const data = await res.json()
+      console.log(data)
 
       if (res.ok) {
-        // Actualiza el estado con los datos actualizados
-        setProducts((prevProducts) =>
-          prevProducts.map((product) =>
-            product.id === updatedProduct.id ? data : product
-          )
-        );
-        console.log("producto actualizado")
-        setIsAddModalOpen(false);
+        await traerProducto();
+        await Swal.fire({
+          position: "center",
+          icon: "success",
+          title: "Producto actualizado",
+          showConfirmButton: false,
+          timer: 1500,
+        });
       } else {
-        console.error("Error al actualizar el producto:", data.message);
-        console.log(updatedProduct)
+        console.error("Error al modificar el producto:", data.message);
+        await Swal.fire({
+          position: "center",
+          icon: "error",
+          title: "Error al actualizar el producto",
+          showConfirmButton: true,
+        });
       }
+      
     } catch (error) {
-      console.error("Error en la solicitud:", error);
-    }
+      console.log(error)
+    } 
   };
 
   const handleEditProduct = (product) => {
@@ -96,30 +111,44 @@ export default function ProductsPage() {
   };
 
   const handleDeleteProduct = async (productId) => {
-    try {
-      const res = await fetch(`http://localhost:3010/producto/${productId}`, {
-        method: "DELETE",
-        headers: {
-          Authorization: `Bearer ${user.token}`,
-        },
-      });
-
-      if (res.ok) {
-        setProducts((prevProducts) =>
-          prevProducts.filter((product) => product.id !== productId)
-        );
-      } else {
-        const errorData = await res.json();
-        console.error("Error al eliminar el producto:", errorData.message);
+    const prodcutToDelete = products.find((product) => product.id === productId);
+    const result = await Swal.fire({
+      title: "¿Estás seguro?",
+      text: `Se eliminará el producto ${prodcutToDelete.name}. Esta acción no se puede deshacer.`,
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Sí, eliminar",
+      cancelButtonText: "Cancelar",
+    });
+    if (result.isConfirmed) {
+      try {
+        const res = await fetch(`http://localhost:3010/producto/${productId}`, {
+          method: "DELETE",
+          headers: {
+            Authorization: `Bearer ${user.token}`,
+          },
+        });
+  
+        if (res.ok) {
+          setProducts((prevProducts) =>
+            prevProducts.filter((product) => product.id !== productId)
+          );
+        } else {
+          const errorData = await res.json();
+          console.error("Error al eliminar el producto:", errorData.message);
+        }
+      } catch (error) {
+        console.error("Error en la solicitud:", error);
       }
-    } catch (error) {
-      console.error("Error en la solicitud:", error);
     }
+    
   };
 
   return (
     <>
-      <div className="user-container">
+      <div className="stock-container p-6  bg-white rounded-md shadow-md">
         <button
           onClick={() => {
             setEditingProduct(null);
